+++ rules/ecmascript-6.js
@@ -15,7 +15,7 @@ module.exports = {
   'constructor-super': 2,
 
   // Enforce the spacing around the * in generator functions
-  'generator-star-spacing': [2, { 'before': false, 'after': true }],
+  'generator-star-spacing': [2, { before: false, after: true }],
 
   // Disallow arrow functions where a condition is expected
   'no-arrow-condition': 2,