+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All changes to this project are documented in this file. This project adheres to
 - Order rules alphabetically
 - Change `eslint` peer dependency to ^1.10.0
 - Split Node specific `env` and `rules` into `ryansobol/node` config
+- Enforce `consistent-as-needed` for this project's `quote-props` rule
 
 #### Best Practices
 - Change `block-scoped-var` rule to an error